--- bin/manage
@@ -1,1 +1,0 @@
-docker exec -it top-assist-app-1 poetry run python main.py
