+++ slack/bot.py
@@ -6,7 +6,6 @@ from credentials import slack_bot_user_oauth_token, slack_app_level_token
 from slack.channel_message_handler import ChannelMessageHandler
 from slack.client import get_bot_user_id
 
-from database.database import get_db_session
 import time
