+++ slack/event_consumer.py
@@ -7,14 +7,14 @@ import logging
 from datetime import datetime
 from pydantic import BaseModel
 from slack_sdk import WebClient
+from configuration import question_context_pages_count
 from credentials import slack_bot_user_oauth_token
-from vector.chroma import retrieve_relevant_documents
-from open_ai.assistants.query_assistant_from_documents import query_assistant_with_context
 from database.interaction_manager import QAInteractionManager
 from database.score_manager import ScoreManager
+from open_ai.assistants.query_assistant_from_documents import query_assistant_with_context
 from database.database import get_db_session
 from slack_sdk.errors import SlackApiError
-
+import vector.pages
 
 class QuestionEvent(BaseModel):
     text: str
@@ -65,7 +65,7 @@ class EventConsumer:
         channel_id = question_event.channel
         message_ts = question_event.ts
         try:
-            context_page_ids = retrieve_relevant_documents(question_event.text)
+            context_page_ids = vector.pages.retrieve_relevant_ids(question_event.text, count=question_context_pages_count)
             response_text, assistant_thread_id = query_assistant_with_context(question_event.text,
                                                                               context_page_ids,
                                                                               self.session,
@@ -113,7 +113,7 @@ class EventConsumer:
         if existing_interaction:
             extended_context_query = self.generate_extended_context_query(existing_interaction, feedback_event.text)
             print(f"\n\nExtended context: {extended_context_query}\n\n")
-            page_ids = retrieve_relevant_documents(extended_context_query)
+            page_ids = vector.pages.retrieve_relevant_ids(extended_context_query, count=question_context_pages_count)
             try:
                 response_text, assistant_thread_id = query_assistant_with_context(feedback_event.text, page_ids, assistant_thread_id)
             except Exception as e:
